Added vcpkg manifest file Changed Shader compilation command to include debugging symbols Added Debug Callback to OpenGL Contexts Added STB as a dependency Added Textures Refactored shaders introspection in Pipeline Added more meaningful errors in Pipeline Added Uniforms Removed unnecessary logs from Shader Added unary operator- to Vector2 and renamed function names to cohere with coding style Fixed Normalize() using Magnitude2() instead of Magnitude() Added mathematical functions Added Vector3 and Vector4 Added Matrix types Added a simple 3d example in Application

+++ src/engine/Linux/Window.cpp
@@ -108,7 +108,10 @@ void Window::SetVSync(bool vsync)
 [[nodiscard]] Vector2<uint32_t> Window::GetSize() const
 {
     if(m_fullscreen)
-        return {DefaultScreenOfDisplay(m_display)->width, DefaultScreenOfDisplay(m_display)->height};
+        return {
+            static_cast<uint32_t>(DefaultScreenOfDisplay(m_display)->width),
+            static_cast<uint32_t>(DefaultScreenOfDisplay(m_display)->height)
+        };
 
     return m_size;
 }
